# Re-introduce double buffer in UpdatePosition, to fix perf regression in gpu_hist