Closes #6552 by partially reverting the commit f779980f7ea7f6f07e86229b8e78144e8a74e6b3