@hcho3 2021-03-17
@trivialfis The posted example in #6552 won't complete in a reasonable amount of time with the latest master (and 1.3.0). I can leave it up over night and see if it can finish by tomorrow.

On the other hand, here's the result with the proposed fix:
```
[0]     convergence-merror:0.10066
[1]     convergence-merror:0.10053
[2]     convergence-merror:0.09998
[3]     convergence-merror:0.09928
[4]     convergence-merror:0.09851
[5]     convergence-merror:0.09770
[6]     convergence-merror:0.09689
[7]     convergence-merror:0.09601
[8]     convergence-merror:0.09517
[9]     convergence-merror:0.09438
Time elapsed = 673.7375219200039 s
```