@trivialfis 2021-03-17
Thanks for sharing what about more conventional dataset?