@trivialfis 2021-03-17
> I can leave it up over night and see if it finishes by tomorrow.

Nah, no need.  I'm just worrying about there might be regression on other types of data.