@hcho3 2021-03-17
> what about more conventional dataset?

Any suggestion? Should I try gbm-bench?