@trivialfis 2021-03-17
> Any suggestion? Should I try gbm-bench?

That would be great!