@hcho3 2021-03-18
@trivialfis Here is the benchmark results on gbm-bench. I do not see any performance degradation:

Runtime (s) | Before | After
-- | -- | --
airline | 40.44 | 40.57
bosch | 2.77 | 2.77
covtype | 4.37 | 4.38
epsilon | 16.12 | 16.19
fraud | 0.44 | 0.41
higgs | 6.50 | 6.53
year | 0.75 | 0.72